pff, float methods actually need double params. still no sound...

+++ softcut~.cpp
@@ -42,69 +42,70 @@ void softcut_dblclick(t_softcut *x);
 
 ///////////
 /// softcut methods
-void softcut_filter_fc(t_softcut *sc, float val) { 
+void softcut_filter_fc(t_softcut *sc, double val) {
     sc->scv.setFilterFc(val);
 }
-void softcut_filter_fc_mod(t_softcut *sc, float val) { 
+void softcut_filter_fc_mod(t_softcut *sc, double val) {
     sc->scv.setFilterFcMod(val);
 }
-void softcut_filter_rq(t_softcut *sc, float val) { 
+void softcut_filter_rq(t_softcut *sc, double val) {
     sc->scv.setFilterRq(val);
 }
-void softcut_filter_lp(t_softcut *sc, float val) { 
+void softcut_filter_lp(t_softcut *sc, double val) {
     sc->scv.setFilterLp(val);
 }
-void softcut_filter_hp(t_softcut *sc, float val) { 
+void softcut_filter_hp(t_softcut *sc, double val) {
     sc->scv.setFilterHp(val);
 }
-void softcut_filter_bp(t_softcut *sc, float val) { 
+void softcut_filter_bp(t_softcut *sc, double val) {
     sc->scv.setFilterBp(val);
 }
-void softcut_filter_br(t_softcut *sc, float val) { 
+void softcut_filter_br(t_softcut *sc, double val) {
     sc->scv.setFilterBr(val);
 }
-void softcut_filter_dry(t_softcut *sc, float val) { 
+void softcut_filter_dry(t_softcut *sc, double val) {
     sc->scv.setFilterDry(val);
 }
-void softcut_rate(t_softcut *sc, float val) {
+void softcut_rate(t_softcut *sc, double val) {
     post("rate %f", val);
     sc->scv.setRate(val);
 }
-void softcut_play(t_softcut *sc, float val) {
+void softcut_play(t_softcut *sc, double val) {
     post("play %f", val);
     sc->scv.setPlayFlag(val > 0.f);
 }
-void softcut_position(t_softcut *sc, float val) { 
+
+void softcut_position(t_softcut *sc, double val) {
     sc->scv.cutToPos(val);
 }
-void softcut_fade_time(t_softcut *sc, float val) { 
+void softcut_fade_time(t_softcut *sc, double val) {
     sc->scv.setFadeTime(val);
 }
-void softcut_loop(t_softcut *sc, float val) { 
+void softcut_loop(t_softcut *sc, double val) {
     sc->scv.setLoopFlag(val);
 }
-void softcut_loop_start(t_softcut *sc, float val) { 
+void softcut_loop_start(t_softcut *sc, double val) {
     sc->scv.setLoopStart(val);
 }
-void softcut_loop_end(t_softcut *sc, float val) { 
+void softcut_loop_end(t_softcut *sc, double val) {
     sc->scv.setLoopEnd(val);
 }
-void softcut_rec(t_softcut *sc, float val) { 
+void softcut_rec(t_softcut *sc, double val) {
     sc->scv.setRecFlag(val > 0.f);
 }
-void softcut_rec_level(t_softcut *sc, float val) { 
+void softcut_rec_level(t_softcut *sc, double val) {
     sc->scv.setRecLevel(val);
 }
-void softcut_pre_level(t_softcut *sc, float val) { 
+void softcut_pre_level(t_softcut *sc, double val) {
     sc->scv.setPreLevel(val);
 }
-void softcut_rec_offset(t_softcut *sc, float val) { 
+void softcut_rec_offset(t_softcut *sc, double val) {
     sc->scv.setRecOffset(val);
 }
-void softcut_level_slew_time(t_softcut *sc, float val) {
+void softcut_level_slew_time(t_softcut *sc, double val) {
     sc->scv.setLevelSlewTime(val);
 }
-void softcut_rate_slew_time(t_softcut *sc, float val) { 
+void softcut_rate_slew_time(t_softcut *sc, double val) {
     sc->scv.setRateSlewTime(val);
 }
